+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -624,6 +624,7 @@ mod tests {
                 sentry_tags: ~,
                 received: ~,
                 measurements: ~,
+                _metrics_summary: ~,
                 other: {},
             },
         ]